Fix conversion warnings when setting alpha to QColor
Adapt to 5bb4baae0379d5903f547f0399be9620f5ab06a0.

+++ src/widgets/itemviews/qitemdelegate.cpp
@@ -1017,7 +1017,7 @@ QPixmap QItemDelegate::selectedPixmap(const QPixmap &pixmap, const QPalette &pal
 
         QColor color = palette.color(enabled ? QPalette::Normal : QPalette::Disabled,
                                      QPalette::Highlight);
-        color.setAlphaF((qreal)0.3);
+        color.setAlphaF(0.3f);
 
         QPainter painter(&img);
         painter.setCompositionMode(QPainter::CompositionMode_SourceAtop);
